# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: ISP Billing Platform API
  description: |
    Multi-tenant SaaS backend for ISP billing, FTTH network management, and TR-069 device provisioning.
    All protected endpoints require a valid JWT Bearer token in the Authorization header.
    Tenant isolation is enforced automatically via JWT claims.
  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Message:
      type: object
      properties:
        message:
          type: string
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...
    Broadcast:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        target: { type: string }
        total_sent: { type: integer }
        total_success: { type: integer }
        total_failed: { type: integer }
        sent_by: { type: string }
        created_at: { type: string, format: date-time }
    BroadcastCreate:
      type: object
      required: [type, title, message]
      properties:
        type: { type: string }
        title: { type: string }
        message: { type: string }
        target: { type: string }
...
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    PerPageParam:
      name: per_page
      in: query
      schema: { type: integer, default: 20 }
    SearchParam:
      name: search
      in: query
      schema: { type: string }
    IDPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
security:
  - BearerAuth: []
paths:
...
  /broadcasts:
    get:
      tags: [Broadcasts]
      summary: List broadcasts
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PerPageParam" }
        - { $ref: "#/components/parameters/SearchParam" }
        - { name: type, in: query, schema: { type: string } }
      responses:
        "200":
          description: Paginated broadcasts
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/PaginationMeta" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Broadcast" }
    post:
      tags: [Broadcasts]
      summary: Send broadcast
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BroadcastCreate" }
      responses:
        "201":
          description: Broadcast sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Broadcast" }
  /broadcasts/{id}:
    get:
      tags: [Broadcasts]
      summary: Get broadcast by ID
      parameters: [{ $ref: "#/components/parameters/IDPath" }]
      responses:
        "200":
          description: Broadcast detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Broadcast" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Broadcasts]
      summary: Delete broadcast
      parameters: [{ $ref: "#/components/parameters/IDPath" }]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }

  # ── Customer Logs ──
...